W6/requirements.txt
dash==2.17.1
plotly==5.22.0
pandas>=2.0.0
pyserial>=3.5
